chore: migrate common sorted set integration tests
