+++ packages/client-sdk-nodejs/test/integration/shared/sorted-set.test.ts
@@ -1,0 +1,6 @@
+import {runSortedSetTests} from '@gomomento/common-integration-tests';
+import {SetupIntegrationTest} from '../integration-setup';
+
+const {Momento, IntegrationTestCacheName} = SetupIntegrationTest();
+
+runSortedSetTests(Momento, IntegrationTestCacheName);
