+++ packages/client-sdk-nodejs/test/integration/topic-client.test.ts
@@ -10,11 +10,13 @@ import {
   InvalidArgumentError,
   SubscribeCallOptions,
 } from '../../src';
-import {ItBehavesLikeItValidatesCacheName} from '@gomomento/common-integration-tests';
+import {
+  ItBehavesLikeItValidatesCacheName,
+  ValidateCacheProps,
+} from '@gomomento/common-integration-tests';
 import {
   SetupIntegrationTest,
   IntegrationTestCacheClientProps,
-  ValidateCacheProps,
 } from './integration-setup';
 import {TextEncoder} from 'util';
 import {
